What Is Mechanum Wheels and How Does It Work?
Ilon's Mecanum wheel is based on the principle of a central wheel with a number of rollers placed at an angle around the periphery of the wheel. The angled peripheral rollers translate a portion of the force in the rotational direction of the wheel to a force normal to the wheel direction.